requirements.txt
uvicorn==0.30.1
fastapi==0.110.3
python-dotenv==1.0.0
crewai==0.51.1
langchain==0.2.15
langchain-aws==0.1.17
langchain-cohere==0.1.5
langchain-community==0.2.15
langchain-core==0.2.37
langchain-openai==0.1.7
langchain-text-splitters==0.2.2
sqlalchemy==2.0.31
tiktoken==0.7.0
pydantic==2.8.2
async-timeout==4.0.3
psycopg2-binary==2.9.9
anthropic
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.12.2
